feat: Add grid with boxes that change to the color gray when hovered over
Added a grid with boxes that change to the color gray when the mouse is hovered over.
When the user hovers their mouse over the on of the cells in the grid the cell changes color.

This was achived by using an event listener, and having the event change the class of the grid.

Closes #1