Add Input field to accept user input to change the number and size of the grid
Changed the number and size of gird using the input field, This was done by using javascript to create the field
then adding an event listner to the field, after adding a keydown event and specifiying the enter key I had the field store
the user's input in a variable to be used for later.

close #2
However after adding this feature there are some performace issues.