Fix Perfomance issue concering user input
Perfomance issue of program concerning user input fixed, moving the ChangeBackground function from inside
of the updateGrid for loop to outside of the loop.

This change significantly increased perfomance.
The problem was that a for each loop was being executed for each div inside of another loop itself. which would
of ocurse increase the amount of computing resources necessary to complete the objective.

close #3